modification: organized folders for every page and component (movies, your list, home, etc)
