+++ my-movie-list/src/App.js
@@ -1,7 +1,7 @@
 import "./App.css";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import React from "react";
-import Home from "./pages/Home";
+import Home from "./pages/home/Home";
 import Login from "./pages/user-auth/Login";
 import Register from "./pages/user-auth/Register";
 import NoPage from "./components/NoPage";
