+++ my-movie-list/src/pages/help/Help.js
@@ -1,3 +1,5 @@
+import "./Help.css";
+
 function Help() {
   return (
     <>
