+++ my-movie-list/src/pages/movies/Movies.js
@@ -1,3 +1,5 @@
+import "./Movies.css";
+
 function Movies() {
   return (
     <>
