+++ my-movie-list/src/pages/user-auth/Login.js
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { useNavigate, Link } from "react-router-dom";
-import "./Login.css";
+import "./Auth.css";
 import axios from "axios";
 import { useAuth } from "../../contexts/AuthContext";
