+++ my-movie-list/src/pages/user-auth/Register.js
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { useNavigate, Link } from "react-router-dom";
 import axios from "axios";
-import "./Login.css";
+import "./Auth.css";
 import { useAuth } from "../../contexts/AuthContext";
 
 function Register() {
