+++ my-movie-list/src/pages/user-list/YourList.js
@@ -1,3 +1,5 @@
+import "YourList.css";
+
 function YourList() {
   return (
     <>
